implement node switch, fix dep cycle
