+++ src/api/index.ts
@@ -4,3 +4,5 @@
 export * from './instantiate';
 export * from './util';
 export * from './contract';
+export * from './loadAccounts';
+export * from './connect';
