+++ src/constants/index.ts
@@ -4,7 +4,10 @@
 import BN from 'bn.js';
 import type { ApiState } from 'types';
 
-export const LOCAL_NODE = 'ws://127.0.0.1:9944'; //wss://rococo-canvas-rpc.polkadot.io
+export enum RPCS {
+  LOCAL = 'ws://127.0.0.1:9944',
+  CANVAS = 'wss://rococo-canvas-rpc.polkadot.io',
+}
 export const DEFAULT_DECIMALS = 12;
 
 export const MAX_CALL_WEIGHT = new BN(2_000_000_000_000);
@@ -19,7 +22,7 @@ export const NULL_CHAIN_PROPERTIES = {
 
 export const INIT_STATE: ApiState = {
   ...NULL_CHAIN_PROPERTIES,
-  endpoint: LOCAL_NODE,
+  endpoint: RPCS.LOCAL,
   keyringStatus: null,
   error: null,
   status: 'CONNECT_INIT',
