+++ src/ui/components/contract/Interact.tsx
@@ -7,17 +7,11 @@ import { ResultsOutput } from './ResultsOutput';
 import { AccountSelect } from 'ui/components/account';
 import { Dropdown, Button, Buttons } from 'ui/components/common';
 import { ArgumentForm, InputGas, InputBalance, Form, FormField } from 'ui/components/form';
-import {
-  createMessageOptions,
-  dryRun,
-  NOOP,
-  prepareContractTx,
-  sendContractQuery,
-  transformUserInput,
-} from 'api';
+import { dryRun, NOOP, prepareContractTx, sendContractQuery, transformUserInput } from 'api';
 import { useApi, useTransactions } from 'ui/contexts';
 import { BN, CallResult, ContractPromise, RegistryError, SubmittableResult } from 'types';
 import { useWeight, useBalance, useArgValues, useFormField, useAccountId } from 'ui/hooks';
+import { createMessageOptions } from 'ui/util/dropdown';
 
 interface Props {
   contract: ContractPromise;
@@ -114,7 +108,7 @@ export const InteractTab = ({ contract }: Props) => {
   const read = async () => {
     const { result, output } = await sendContractQuery(
       contract.query[message.method],
-      keyring?.getPair(accountId),
+      keyring.getPair(accountId),
       options,
       transformed
     );
