+++ src/ui/util/index.ts
@@ -13,13 +13,6 @@ export function truncate(value: string | undefined, sideLength = 6): string {
     : '';
 }
 
-export const MOCK_CONTRACT_DATA: [string, number, string[], unknown[]][] = [
-  ['Flipper', 0, ['alpha', 'beta'], [true]],
-  ['ERC20', 1, ['alpha', 'beta', 'gamma'], [1000000000000n * 1000000n]],
-  ['DNS', 2, ['delta'], []],
-  ['Incrementer', 1, ['beta', 'delta', 'gamma'], [7]],
-];
-
 export function isValidCodeHash(value: string): boolean {
   return /^0x[0-9a-fA-F]{64}$/.test(value);
 }
